Correção de imports e reorganização dos repositórios para o pacote co…

+++ src/main/java/br/com/sigest/tesouraria/domain/entity/ContaFinanceira.java
@@ -3,6 +3,7 @@ package br.com.sigest.tesouraria.domain.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 /**
  * Entidade que representa uma Conta Financeira.
@@ -20,5 +21,5 @@ public class ContaFinanceira {
     private String nome;
 
     @Column(nullable = false)
-    private Float saldoAtual;
+    private BigDecimal saldoAtual;
 }
